requirements.txt
pandas
sentence-transformers
openpyxl
duckdb
numpy
requests
flask
openai
python-dotenv
atlassian-python-api
